requirements.txt
qwen-tts==0.1.1
faster-whisper==1.2.0
openai-whisper==20250625
gradio==5.50.0
librosa==0.11.0
numpy==2.4.3
soundfile==0.13.1
